New clamour analysis

Creates a new analysis file based on the clamour template.

clamour_new(
  name,
  hashtag,
  description = paste("An analysis of", hashtag),
  start_day = Sys.Date(),
  end_day = Sys.Date() + 1,
  timezone = Sys.timezone(),
  theme = "theme_light",
  accent = "dodgerblue",
  accent2 = NULL,
  kcore = 2,
  topics_k = 6,
  bigram_filter = 3,
  fixed = FALSE,
  seed = 1,
  dir = "analysis"
)

Arguments

name
Name of the analysis. This is used to name files so should usually be short and simple.

hashtag
The hashtag to use for this analysis, starting with "#". Can also be a vector of hashtags.

description
A more detailed description of the analysis, details of the event etc.

start_day
First day to be included in the analysis.

end_day
Last day to be included in the analysis.

timezone
Timezone to be used for the analysis.

theme
Name of a ggplot2 theme function to use for plots.

accent
Accent colour to use for plots.

accent2
Secondary accent colour to use for plots. If NULL this will be created based on accent.

kcore
k-core to use for filtering the mentions network.

topics_k
Number of topics to use for topic modelling.

bigram_filter
Number of co-occurrences for filtering the bigram graph.

fixed
Whether to continue to search and download new tweets. Set to TRUE to fix the dataset used for analysis.

seed
Random seed set at the start of analysis.

dir
Directory to save the new analysis file.

Value

Path to the newly created file, invisibly