Setup a clamour project

This function creates the directory structure necessary for analysing Twitter hashtag activity and displaying the results on a website. It will also prompt you to install packages required by the default analysis and to authorise rtweet.

clamour_setup(path, rstudio = rstudioapi::isAvailable(), open = interactive())

Arguments

path
A path for the project. If it exists it will be used. If it does exist it will be created, provided the parent path exists.

rstudio
If TRUE, calls usethis::use_rstudio() to make the new project into an RStudio Project. If FALSE a sentinel .here file is placed so that the directory can be recognized as a project by the here or rprojroot packages.

open
If TRUE, activates the new project:

If using RStudio desktop, the package is opened in a new session.

If on RStudio server, the current RStudio project is activated.

Otherwise, the working directory and active project is changed.

Value

Path to the newly created project, invisibly.

References

Based on usethis::create_package()